requirements.txt
Flask==3.1.1
nltk==3.9.1
openai==1.82.0
pdfplumber==0.11.6
python-dotenv==1.1.0
Requests==2.32.3
Werkzeug==3.1.3
